Extract `VmmMode` to a separate file out of `cuttlefish_config.h`
Bug: b/435249938
