+++ base/cvd/cuttlefish/host/commands/assemble_cvd/BUILD.bazel
@@ -161,6 +161,7 @@ cf_cc_library(
         "//cuttlefish/host/libs/config:data_image",
         "//cuttlefish/host/libs/config:fetcher_config",
         "//cuttlefish/host/libs/config:instance_nums",
+        "//cuttlefish/host/libs/config:vmm_mode",
         "//cuttlefish/host/libs/feature:inject",
         "//libbase",
         "@gflags",
@@ -174,7 +175,7 @@ cf_cc_library(
     deps = [
         "//cuttlefish/common/libs/utils:files",
         "//cuttlefish/common/libs/utils:result",
-        "//cuttlefish/host/libs/config:cuttlefish_config",
+        "//cuttlefish/host/libs/config:vmm_mode",
         "//cuttlefish/host/libs/image_aggregator",
         "//cuttlefish/host/libs/image_aggregator:qcow2",
         "//libbase",
@@ -196,6 +197,7 @@ cf_cc_library(
         "//cuttlefish/host/libs/config:cuttlefish_config",
         "//cuttlefish/host/libs/config:fetcher_config",
         "//cuttlefish/host/libs/config:instance_nums",
+        "//cuttlefish/host/libs/config:vmm_mode",
         "//libbase",
     ],
 )
@@ -267,6 +269,7 @@ cf_cc_library(
         "//cuttlefish/host/libs/config:host_tools_version",
         "//cuttlefish/host/libs/config:instance_nums",
         "//cuttlefish/host/libs/config:secure_hals",
+        "//cuttlefish/host/libs/config:vmm_mode",
         "//cuttlefish/host/libs/vhal_proxy_server",
         "//cuttlefish/host/libs/vm_manager",
         "//libbase",
@@ -326,6 +329,7 @@ cf_cc_library(
         "//cuttlefish/host/graphics_detector:graphics_detector_cc_proto",
         "//cuttlefish/host/libs/config:config_constants",
         "//cuttlefish/host/libs/config:cuttlefish_config",
+        "//cuttlefish/host/libs/config:vmm_mode",
         "//libbase",
         "@fmt",
         "@protobuf",
