+++ base/cvd/cuttlefish/host/commands/assemble_cvd/disk/BUILD.bazel
@@ -195,6 +195,7 @@ cf_cc_library(
         "//cuttlefish/host/commands/assemble_cvd/disk:generate_persistent_vbmeta",
         "//cuttlefish/host/libs/config:ap_boot_flow",
         "//cuttlefish/host/libs/config:cuttlefish_config",
+        "//cuttlefish/host/libs/config:vmm_mode",
         "//cuttlefish/host/libs/image_aggregator",
         "//cuttlefish/host/libs/vm_manager",
         "@gflags",
@@ -308,6 +309,7 @@ cf_cc_library(
         "//cuttlefish/common/libs/utils:result",
         "//cuttlefish/host/libs/config:cuttlefish_config",
         "//cuttlefish/host/libs/config:data_image",
+        "//cuttlefish/host/libs/config:vmm_mode",
         "//cuttlefish/host/libs/image_aggregator:qcow2",
     ],
 )
