+++ base/cvd/cuttlefish/host/commands/assemble_cvd/disk/sd_card.cc
@@ -20,6 +20,7 @@
 #include "cuttlefish/common/libs/utils/result.h"
 #include "cuttlefish/host/libs/config/cuttlefish_config.h"
 #include "cuttlefish/host/libs/config/data_image.h"
+#include "cuttlefish/host/libs/config/vmm_mode.h"
 #include "cuttlefish/host/libs/image_aggregator/qcow2.h"
 
 namespace cuttlefish {
