+++ base/cvd/cuttlefish/host/commands/assemble_cvd/disk_builder.cpp
@@ -22,7 +22,7 @@
 #include <android-base/file.h>
 
 #include "cuttlefish/common/libs/utils/files.h"
-#include "cuttlefish/host/libs/config/cuttlefish_config.h"
+#include "cuttlefish/host/libs/config/vmm_mode.h"
 #include "cuttlefish/host/libs/image_aggregator/image_aggregator.h"
 #include "cuttlefish/host/libs/image_aggregator/qcow2.h"
