+++ base/cvd/cuttlefish/host/commands/assemble_cvd/disk_builder.h
@@ -19,7 +19,7 @@
 #include <vector>
 
 #include "cuttlefish/common/libs/utils/result.h"
-#include "cuttlefish/host/libs/config/cuttlefish_config.h"
+#include "cuttlefish/host/libs/config/vmm_mode.h"
 #include "cuttlefish/host/libs/image_aggregator/image_aggregator.h"
 
 namespace cuttlefish {
