+++ base/cvd/cuttlefish/host/commands/assemble_cvd/flags/BUILD.bazel
@@ -90,7 +90,7 @@ cf_cc_library(
         "//cuttlefish/common/libs/utils:result",
         "//cuttlefish/host/commands/assemble_cvd:flags_defaults",
         "//cuttlefish/host/commands/assemble_cvd:guest_config",
-        "//cuttlefish/host/libs/config:cuttlefish_config",
+        "//cuttlefish/host/libs/config:vmm_mode",
         "//libbase",
         "@gflags",
     ],
