+++ base/cvd/cuttlefish/host/commands/metrics/BUILD.bazel
@@ -21,6 +21,7 @@ cf_cc_binary(
         "//cuttlefish/common/libs/utils:tee_logging",
         "//cuttlefish/host/commands/metrics/proto:metrics_protos",
         "//cuttlefish/host/libs/config:cuttlefish_config",
+        "//cuttlefish/host/libs/config:vmm_mode",
         "//cuttlefish/host/libs/metrics",
         "//cuttlefish/host/libs/metrics:metrics_headers",
         "//cuttlefish/host/libs/msg_queue",
